pyproject.toml
[project]
name = "coderag-rag-mcp"
version = "0.1.0"
description = "Code RAG MCP Server for code retrieval"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "weaviate-client>=3.26.2",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "llama-index-vector-stores-weaviate>=0.1.0",
    "mcp>=0.9.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.52.2",
    "llama-index-llms-ollama>=0.9.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

